:rainbow: move class inheritance to separate files
