+++ website_multi_company_sale/__manifest__.py
@@ -4,7 +4,7 @@
     "category": "eCommerce",
     "live_test_url": "http://apps.it-projects.info/shop/product/website-multi-company?version=11.0",
     "images": ["images/website_multi_company_sale_main.png"],
-    "version": "11.0.1.5.0",
+    "version": "11.0.1.5.1",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",
