+++ website_multi_company_sale/doc/changelog.rst
@@ -1,3 +1,7 @@
+`1.5.1`
+-------
+- **Improvement:** better organized code
+
 `1.5.0`
 -------
 - **New:** List of available countries on billing address page is configurable per website
