+++ website_multi_company_sale/models/res_partner.py
@@ -1,0 +1,7 @@
+from odoo import models, fields
+
+
+class ResPartner(models.Model):
+    _inherit = 'res.partner'
+
+    last_website_so_id = fields.Many2one(company_dependent=True, website_dependent=True)
